Fix `_id` and `_score` in records APIs (#334)
## Problem
Code generation from OpenAPI specs has some quirks, one of which is
trying to get rid of leading underscores in objects and classes in some
languages. We currently use `_id` and `_score` in our records API to
denote reserved keys. We previously needed to add some logic during
codegen in python to prevent this same issue.

Currently:
- You cannot call `upsertRecords` with records containing `_id`.
- When calling `searchRecords` the resulting object has the keys mapped
from `_id` -> `id` and `_score` -> `score`.

## Solution
- Updated `IntegratedRecord` type which resolves the problem with
`upsertRecords`
- Add a new function to the `build-oas.sh` script that handles massaging
the `Hit.ts` file fields to maintain the shape of the API.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI Unit + Integration