Allow `embed` as a parameter to `configureIndex` (#338)
## Problem
Currently, the `ConfigureIndexRequest` interface supports `embed?:
ConfigureIndexRequestEmbed` as a parameter, which aligns with the OAS
spec and how the endpoint can be used:
https://github.com/pinecone-io/pinecone-ts-client/blob/e5b39d7a6c77836b5b0ae1f696ce7e6f6020f82e/src/pinecone-generated-ts-fetch/db_control/models/ConfigureIndexRequest.ts#L64

However, the `embed` field was not added to
`ConfigureIndexRequestProperties`, defining what we expect a user to
provide in a request. We also error out if the user attempts to
configure an index with just the `embed` property.

## Solution
Update `ConfigureIndexRequestProperties` and the validation logic in the
`configureIndex` action to allow leveraging `embed` properly.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI Unit & Integration Tests