Expose `includeUrl` in `Assistant.describeFile`, add `sourceCollection` for serverless (#349)
## Problem
We got feedback that currently, the `Assistant.describeFile` method does
not allow passing `includeUrl` along with `fileId`. It's currently
hard-coded to `true`.

`2025-04` also supports `sourceCollection` as a field for
`ServerlessSpec` as a means of migrating from pods-based to serverless
indexes.

## Solution
- Regenerate core from latest `2025-04` spec.
- Update `CreateIndexServerlessSpec` to include `sourceCollection`.
Other types are exported from generated code, and already include the
field.
- Update `Assistant.describeFile` method to allow `includeUrl` as an
optional argument. Currently, I've left it defaulting to `true` as that
was the hard-coded behavior previously, and I don't want to break anyone
who was relying on this.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI - unit, integration, and external app tests