Add `tsx` as a `devDependency`, clean up `setup` GH action (#352)
## Problem
Since we have scripts like `/scripts/test-cleanup.ts` that are written
in typescript and run as a part of CI workflows, we should really be
using `tsx` as a direct `devDependency` so it's pulled in as a part of
`npm install`.

After updating how we run a couple scripts in `package.json`, I noticed
the cron job for resource cleanup is now failing for a different reason:
https://github.com/pinecone-io/pinecone-ts-client/actions/runs/17567893284/job/49898351307

## Solution

- Add `tsx` as an explicit `devDependency`.
- Clean up the `/actions/testing-cleanup.yml` action to remove the
unneeded pinecone_api_key input, and properly default to node `v18.x` if
the version isn't provided.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Merge and make sure the cron job passes later this evening.